+++ src/main/java/mssql/googlecode/concurrentlinkedhashmap/LinkedDeque.java
@@ -434,27 +434,3 @@ final class LinkedDeque<E extends Linked<E>> extends AbstractCollection<E> imple
     abstract E computeNext();
   }
 }
-
-/**
- * An element that is linked on the {@link Deque}.
- */
-interface Linked<T extends Linked<T>> {
-
-  /**
-   * Retrieves the previous element or <tt>null</tt> if either the element is
-   * unlinked or the first element on the deque.
-   */
-  T getPrevious();
-
-  /** Sets the previous element or <tt>null</tt> if there is no link. */
-  void setPrevious(T prev);
-
-  /**
-   * Retrieves the next element or <tt>null</tt> if either the element is
-   * unlinked or the last element on the deque.
-   */
-  T getNext();
-
-  /** Sets the next element or <tt>null</tt> if there is no link. */
-  void setNext(T next);
-}
